Rename Front directory to UI
