+++ OpenDBDiff/UI/DataCompareForm.resx
