+++ OpenDBDiff/UI/DatabaseProgressControl.cs
@@ -1,6 +1,6 @@
 ﻿using System.Windows.Forms;
 
-namespace OpenDBDiff.Front
+namespace OpenDBDiff.UI
 {
     public partial class DatabaseProgressControl : UserControl
     {
