+++ OpenDBDiff/UI/DatabaseProgressControl.resx
