+++ OpenDBDiff/UI/ErrorForm.resx
