+++ OpenDBDiff/UI/ListProjectsForm.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Windows.Forms;
 
-namespace OpenDBDiff.Front
+namespace OpenDBDiff.UI
 {
     public delegate void ListProjectHandler(Project itemSelected);
