+++ OpenDBDiff/UI/MainForm.resx
