+++ OpenDBDiff/UI/OptionForm.resx
