+++ OpenDBDiff/UI/ProgressForm.resx
